updated multiple objects, added missing Pojos, fixed incorrect types, moved some packages around for consistency.  sql file dump is attached.
git-svn-id: https://svn.kuali.org/repos/kpme/trunk@211 6c5f4195-72ea-6250-e7e4-e498f552a5df

+++ src/main/java/org/kuali/hr/time/clocklog/ClockLog.java
@@ -17,8 +17,8 @@ public class ClockLog extends PersistableBusinessObjectBase {
     private Long clockLogId = null;
     private String principalId;
     private Integer jobNumber;
-    private Integer workAreaId;
-    private Integer taskId;
+    private Long workAreaId;
+    private Long taskId;
     private Calendar clockTimestamp;
     private String clockTimestampTimezone;
     private String clockAction;
@@ -50,19 +50,19 @@ public class ClockLog extends PersistableBusinessObjectBase {
         this.jobNumber = jobNumber;
     }
 
-    public Integer getWorkAreaId() {
+    public Long getWorkAreaId() {
         return workAreaId;
     }
 
-    public void setWorkAreaId(Integer workAreaId) {
+    public void setWorkAreaId(Long workAreaId) {
         this.workAreaId = workAreaId;
     }
 
-    public Integer getTaskId() {
+    public Long getTaskId() {
         return taskId;
     }
 
-    public void setTaskId(Integer taskId) {
+    public void setTaskId(Long taskId) {
         this.taskId = taskId;
     }
